requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=h7f98852_5
brotli-bin=1.0.9=h7f98852_5
brotlipy=0.7.0=py39h3811e60_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cartopy=0.18.0=py39h0d9ca2b_1
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
configparser=5.2.0=pyhd8ed1ab_0
cryptography=35.0.0=py39hbca0aa6_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=he372182_0
docker-pycreds=0.4.0=py_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
geoopt=0.4.0=pypi_0
geos=3.8.0=he6710b0_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.26=pyhd8ed1ab_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=hf484d3e_1000
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=py39h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libbrotlicommon=1.0.9=h7f98852_5
libbrotlidec=1.0.9=h7f98852_5
libbrotlienc=1.0.9=h7f98852_5
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.4.3=py39hf3d152e_2
matplotlib-base=3.4.3=py39hbbc1b5f_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h295c915_0
pathtools=0.1.2=py_1
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
proj=6.2.1=hc80f0dc_0
promise=2.3=py39hf3d152e_5
protobuf=3.15.8=py39he80948d_0
psutil=5.8.0=py39h27cfd23_1
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyqt=5.9.2=py39h2531618_6
pyshp=2.2.0=pyhd8ed1ab_0
pysocks=1.7.1=py39hf3d152e_4
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=py39h3811e60_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
scipy=1.7.3=py39hc147768_0
sentry-sdk=1.5.5=pyhd8ed1ab_0
setuptools=58.0.4=py39h06a4308_0
shapely=1.7.1=py39h1728cc4_0
shortuuid=1.0.8=py39hf3d152e_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.37.2=hc218d9a_0
subprocess32=3.5.4=py_1
termcolor=1.1.0=py_2
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.2=py39_cu102
torchvision=0.11.3=py39_cu102
tornado=6.1=py39h3811e60_1
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd8ed1ab_1
wandb=0.12.10=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yaspin=2.1.0=pyhd8ed1ab_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
